Remove new? since its meaning is ill-defined.
Note that new? would never return true within #save, as of the state of
this code, even for brand new drafts that are being saved as part of a
publish. This is because published drafts come via #publish!, which
calls title=, which sets a value for @source if one exists. Thus, new?
would be false, even though the draft is in some sense new. This is what
is meant by ill-defined.

Fixes #16.