Generate bookmarks with JavaScript.
The bookmark URLs are based on `request.base_url`, but with SSL
configurations, this can erroneously be set to `localhost:4567`. This
means the user has to forward magic HTTP headers correctly, and even
then the port number seems to be broken.

Relying on `window.location` means the URL base will match wherever the
admin server is running.

Fixes #52.