Remove the admin UI and clean up a ton of technical debt.
The admin UI had a bunch of problems and was never going to be the thing
I wanted it to be. It also caused more deployment headaches because now
you have to keep the admin app running. I'm cutting things down to focus
on just the simpler use case of ssh with rsync, etc.

Ordinarily I'd have kept all these commits logical and separate, but
there was so much to rip out that I got too far along. Didn't seem worth
it.